refactor: moved contextTypes to Types directory
relates #236
